+++ tests/test_filestore.py
@@ -1,9 +1,11 @@
 import os.path
+import shutil
 import struct
 import tempfile
 from pathlib import Path
+from unittest import TestCase
 
-from pyfakefs.fake_filesystem_unittest import TestCase
+from spacepackets.cfdp import ChecksumType
 
 from cfdppy.crc import calc_modular_checksum
 from cfdppy.filestore import FilestoreResult, NativeFilestore
@@ -31,8 +33,7 @@ EXAMPLE_DATA_CFDP = bytes(
 
 class TestCfdpHostFilestore(TestCase):
     def setUp(self):
-        self.setUpPyfakefs()
-        self.temp_dir = tempfile.gettempdir()
+        self.temp_dir = tempfile.mkdtemp()
         self.test_file_name_0 = Path(f"{self.temp_dir}/cfdp_unittest0.txt")
         self.test_file_name_1 = Path(f"{self.temp_dir}/cfdp_unittest1.txt")
         self.test_dir_name_0 = Path(f"{self.temp_dir}/cfdp_test_folder0")
@@ -40,7 +41,7 @@ class TestCfdpHostFilestore(TestCase):
         self.test_list_dir_name = Path(f"{self.temp_dir}/list-dir-test.txt")
         self.filestore = NativeFilestore()
 
-        self.file_path = Path(f"{tempfile.gettempdir()}/crc_file")
+        self.file_path = Path(f"{self.temp_dir}/crc_file")
         with open(self.file_path, "wb") as file:
             file.write(EXAMPLE_DATA_CFDP)
         # Kind of re-writing the modular checksum impl here which we are trying to test, but the
@@ -62,6 +63,9 @@ class TestCfdpHostFilestore(TestCase):
         full_sum %= 2**32
 
         self.expected_checksum_for_example = struct.pack("!I", full_sum)
+
+    def tearDown(self):
+        shutil.rmtree(self.temp_dir)
 
     def test_creation(self):
         res = self.filestore.create_file(self.test_file_name_0)
@@ -140,6 +144,11 @@ class TestCfdpHostFilestore(TestCase):
     def test_modular_checksum(self):
         self.assertEqual(calc_modular_checksum(self.file_path), self.expected_checksum_for_example)
 
-    def tearDown(self):
-        if self.file_path.exists():
-            os.remove(self.file_path)
+    def test_zero_length_checksum(self):
+        with self.assertRaises(ValueError):
+            self.filestore.calculate_checksum(
+                checksum_type=ChecksumType.CRC_32,
+                file_path=self.file_path,
+                size_to_verify=10,
+                segment_len=0,
+            )